Allow aspect-ratio to be defined as prop
Currently the aspect ratio is auto detected for the (main) image. But
the placeholder image will always be displayed with a default aspect
ratio. I assume this is because the placeholder can be an image with
different aspect ratio of the main image.

This change will allow the aspect ratio to be explicitly defined with
a prop. If its not defined, the behavior is as before.